Fix SerialPIO sampling, avoid reading random garbage
Adjust the 1/2 bit time to match the number of extra cycles in the actual
PIO loop.

Shift out the entire start bit, which results in sampling the data at the
midpoint and not the starting time of a bit.

Tested at 300bps all the way to 2,000,000bps using a loopback connection.

Fies #360