Add printf_float protection
For reasons not yet fully understood, the Newlib _printf_float is not
multicore safe, even though all locks are implemented in the core.
As a brute force way of fixing this, wrap the printf_float (which is
only called on formatting a floating point) and mutex protect it.

Failing code w/o this patch:
````
void setup() {
  Serial.begin(115200);
}

void loop() {
  static uint64_t cnt = 0;
  int sz = rand() & 127;
  void *ptr = malloc(sz);
  memset(ptr, 0, sz);
  ptr = realloc(ptr, sz + (rand() & 127));
  snprintf((char *)ptr, sz, "%0.12f", cnt / 1.2e12);
  free(ptr);
  cnt++;
  if (! (cnt % 1000)) Serial.printf("Core 0: %d\n", cnt);
}

void setup1() {
}

void loop1() {
  static uint64_t cnt = 0;
  int sz = rand() & 127;
  void *ptr = malloc(sz);
  memset(ptr, 1, sz);
  ptr = realloc(ptr, sz + (rand() & 127));
  snprintf((char *)ptr, sz, "%0.12f", cnt / 1.2e12);
  free(ptr);
  cnt++;
  if (! (cnt % 1000)) Serial.printf("Core 1: %ld\n", cnt);
}
````