Make USB, UART, SPI, and I2C objects non-copyable
Objects that correspond to a single physical device shouldn't be copyable
as their states could end up becoming completely confused if the copy
changes them and the main global object is used later on.

Avoid it by removing the copy and copy move constructors of the
SPI, SerialUSB, SerialUART, and Wire objects.

This mirrors the Arduino MBED core (which uses template NonCopyable<>)