Disable interrupts while in the memory manager
Fixes random crashes while on the PicoW with WiFi.

The system malloc/etc. is not re-entrant, so it is not safe to do
any memory (de)allocation from an interrupt.

The LWIP stack is configured not to use the memory manager, but it may
end up calling callback functions in userspace which may well use
new or malloc() (i.e. to manage a new connection info or process a
UDP packet like in MDNS).

This could cause unpredicatable crashes due to memory corruption at
some random time after the interrupt/malloc occurred.

Avoid the issue completely by locking interrupts out while in the memory
manager.  Inter-core will be protected by the malloc_lock already in
Newlib.

This should not have any effect on absolute performance of the system.