Enable optional GPROF profiling of apps
Apps can now use gprof to analyze performance and runtimes.  This requires
significant RAM and really only makes sense on devices with PSRAM to store
the state.

When the menu item is selected, allocates RAM and tracks function calls and
periodically samples the PC to generate a histogram of application usage.
Data written out to a gmon.out file, which can then be transferred to
a PC using gdb or a SD card.