Build LWIP at app compile time, no need for multi
We needed about 6 copies of the LWIP/BT library for each CPU when the
lwip library was built as part of the SDK libpico.

Remove LWIP from the SDK builds and replace it with the actual LWIP
source (linked via `#include`) so we can adjust for IPv4/IPv^/32K/16K
at runtime.  Shrinks download size of a release significantly.

Still need separate LWIP-BT library because the CYW43 driver is
different when BT is enabled.

Update and clean up the libpico CMakefile and common platform_def.txt

Link to config includes, don't copy, where possible