Handke SPISlave with no output data ever written
Fixes #3068.  The SPISlave TX-FIFO IRQ would infinitely try to fill the FIFO
as soon as IRQs were enabled, but in the case when the app never sent any
data this would cause an infinite loop in the IRQ handler.

Now abort if we don't have any data, even if there is space in the FIFO.