Fix polling and IRQ memory corruption
The IRQ callback routine ended up passing in a stack variable address.  For main
app code this is legal because the app blocks until the LWIP call returns.  But
for IRQs, it returns immediately (can't block in IRQ) and the stack pointer we
passed in will be corrupt.  Use a dumb static (heap) variable for now.

W5100 is now running with multiple AdvancedWebServer clients in parallel with
a WiFiClient MOTD process on core 1, in parallel.

Rewriting the CYW43 driver to work with this TBD.