Fix race condition in WiFiServer
It is possible under heavy load with multithreading that a connection gets a
tcp_abort just before tcp_accept will be called.  When we abort, we clear the
this pointer, and so when we try and recover our object we crash.

Check for aborted connections (and ignore) on WiFiServer, like we do in
WiFiClient