BREAKING: USB PID will vary depending on devices
The USB PID was already being changed depending on the devices installed
in the app.  This was needed because Windows has issues with the same
VID:PID exporting different HID devices (i.e. it would not recognize
changed ones properly).

Unfortunately the change was to set a bit to 1 in the PID, but if the
bit was already 1 then no change would occur.

Now, actually XOR the bitmap of device exports to ensure the VID:PID
of a sketch with only Serial will differ from one with Serial and
Keyboard, for example.