Build USB descriptor from data structures, not code (#3107)
The USB descriptor and HID report descriptors were built in a hard-coded
mess of code with lots of very brittle logic to ensure things lined up
and were the right size.  Weak functions were used to signal the core
that different descriptors needed to be added, so things like the Keyboard
library actually had changes done in the core proper.

Move to a saner and more maintainable, global initialization stage process
where global object descriptors can call a `register` function to plug
in their own USB or HID descriptors to the core process.