Support dynamic USB reconfiguration (#3139)
Now that the USB descriptor and HID reports are handled by data, it is
possible to change then on-the-fly from an application and customize
the USB vendor/product/etc.

It is also possible to add and remove HID and other devices, at runtime.

Serial (USB) now properly handles `Serial.end()` and completely removes
the USB CDC descriptor for the Serial port when called.  That makes
it possible to build sketches which only export HID (Keyboard, Joystick,
Mouse) or MSC (FatFSUSB or SingleFileDrive) devices and no Serial
port.

All USB devices now register themselves on `xxx.begin()` and not
at the start of the sketch.  This could be breaking for sketches
which (incorrectly) forget to call `Keyboard.begin()` before running,
for example.  In most cases this will be a non-issue because most
scripts (all examples here!) so call begin() as appropriate.

Updates the documentation with the new USB calls for end users.
Advanced users can check out the shipped libraries for more
complicated customization (adding/removing custom devices).

Clean up and make the USB subsystem design into a class instead of
a collection of C functions which share static local state.

Also fix Picotool reset compilation issue

* Move all HID_xxx to main/master

Avoid using random branches for the libraries, pull in the changes
for RP2040 onto the master branch of each.