Reduce LWIP flash 5.4K by cleaning debug prints
In non-debug cases the string to `panic()` isn't actually sent to the
Serial port, or anywhere in fact.

When there is no debug port defined, replace the panic(str) of
LWIP_PLATFORM_ASSERT with panic("lwip") instead.  This doesn't
effect debugging because the source line of the LWIP_ASSERT
will still show the full string (GDB shows the source).

Also, LWIP_DEBUG is a flag by definition, not value, so fix
lwipopts.h to not define it or the LWIP_X debug defines unless
LWIP_DEBUG is set by the compliation.

Saves 5480 bytes of flash for LWIP (Pico W, 2W, Ethernet) apps.