-O2 for Pico2 builds, -O3 shows a GCC error
When -O3 is used to build LWIP, GCC emits a bogus error while compiling
DHCP.C:
>  ../dhcp.c:1453:30: error: writing 1 byte into a region of size 0

DHCP_OPTIONS_LEN was verified as being DHCP_MIN_OPTIONS_LEN (68U) and not
0 using #ifs and #errors in the dhcp.h header, but GCC somehow forgets
this while doing it's checks on O3.  Looks to be a bug in GCC to me,
so drop to -O2 for RP204 builds.